snp-bioinfo-service/requirements.txt
# FastAPI y servidor
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Redis para cola de trabajos
redis>=5.2.0
upstash-redis>=1.0.0

# HTTP client async
httpx>=0.28.0

# Bioinformatica
biopython>=1.84

# Validacion y configuracion
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Supabase client
supabase>=2.10.0

# Variables de entorno
python-dotenv>=1.0.1

# Logging estructurado
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
